Refactor benchmarking module and add tests
Rebuilds the benchmark module to replace the criterion_group and
criterion_main macros with a custom main module that allows individual
benchmarks to be setup and ran without compiling the code for the other
ones. As well, the number of iterations and an optional data path can
now be specified via environment variables. Finally, benchmarks for push
and workspace_add were added
