+++ oxen-rust/Cargo.toml
@@ -187,8 +187,8 @@ criterion = { version = "2.10.1", package = "codspeed-criterion-compat", feature
 serial_test = "3.2.0"
 
 [[bench]]
-name = "oxen"
-path = "src/benches/oxen.rs"
+name = "benchmark"
+path = "src/benches/benchmark.rs"
 harness = false
 
 [package.metadata.cargo-machete]
