+++ oxen-rust/src/cli/src/cmd/remote_mode/restore.rs
@@ -1,5 +1,3 @@
-// TODO: Later
-
 use async_trait::async_trait;
 use clap::Command;
